Lab 5 changes done

+++ AED_Lab5/src/Business/Order.java
@@ -46,7 +46,7 @@ public class Order {
         Order.count = count;
     }
     
-    public OrderItem addOrderItem(Product p,int q,int price){
+    public OrderItem addOrderItem(Product p,int q,double price){
         OrderItem o = new OrderItem();
         o.setProduct(p);
         o.setQuantity(q);
